+++ features/de.bund.bfr.knime.fsklab.feature/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Order of model parameters on model import with the creator node.
+
 ## [1.4.5] - 2019-02-13
 
 ### Fixed